requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
allure-pytest==2.14.2
    # via -r requirements.in
allure-python-commons==2.14.2
    # via
    #   -r requirements.in
    #   allure-pytest
attrs==25.3.0
    # via allure-python-commons
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
colorama==0.4.6
    # via pytest
execnet==2.1.1
    # via pytest-xdist
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via pytest-html
markupsafe==3.0.2
    # via jinja2
numpy==2.2.6
    # via -r requirements.in
packaging==25.0
    # via pytest
PyYAML==6.0.1
    # via
    #   allure-python-commons
    #   pytest
pytest==8.3.5
    # via
    #   -r requirements.in
    #   allure-pytest
    #   pytest-html
    #   pytest-metadata
    #   pytest-ordering
    #   pytest-xdist
pytest-html==4.1.1
    # via -r requirements.in
pytest-metadata==3.1.1
    # via pytest-html
pytest-ordering==0.6
    # via -r requirements.in
pytest-xdist==3.7.0
    # via -r requirements.in
python-dotenv==1.1.0
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
urllib3==2.5.0
    # via requests